% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in app_ui.R
\name{add_external_resources}
\alias{add_external_resources}
\title{Add external Resources to the Application}
\usage{
add_external_resources()
}

\description{
This function is internally used to add external
 resources inside the application.
}

