% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in app_server.R
\name{app_server}
\alias{app_server}
\title{The application server-side}
\usage{
app_server(input, output, session)
}

\arguments{
\item{input}{Internal parameters. DO NOT REMOVE.}

\item{output}{Internal parameters. DO NOT REMOVE.}

\item{session}{Internal parameters. DO NOT REMOVE.}
}

\description{
The application server-side
}

