% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in app_config.R
\name{app_sys}
\alias{app_sys}
\title{Access files in the current app}
\usage{
app_sys(...)
}

\arguments{
\item{...}{character vectors, specifying subdirectory and file(s)
 within your package. The default, none, returns the root of the app.}
}

\description{
If you manually change your package name in the DESCRIPTION,
 don't forget to change it here too, and in the config file.
}

