% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{available_formats}
\alias{available_formats}
\title{Available Server parameters expressed as functions}
\usage{
available_formats()
}

\description{
Available Server parameters expressed as functions
}

