% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{ensure_cols}
\alias{ensure_cols}
\title{Ensure data frame contains expected columns to avoid mutate/group_by errors}
\usage{
ensure_cols(df, cols)
}

\arguments{
\item{df}{input data frame}

\item{cols}{vector of expected column names}
}

\description{
Ensure data frame contains expected columns to avoid mutate/group_by errors
}

