% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{fetch_sanctions}
\alias{fetch_sanctions}
\title{Fetch GSDB dyadic sanctions imposed against a country in a year range}
\usage{
fetch_sanctions(con, code, years)
}

\arguments{
\item{con}{SQL connection}

\item{code}{Dynamic code of the sanctioned country}

\item{years}{Numeric vector of years to filter on (min/max are used)}
}

\description{
Fetch GSDB dyadic sanctions imposed against a country in a year range
}

