% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{format_list_and}
\alias{format_list_and}
\title{Join a vector of items into a human-readable "a, b and c" list}
\usage{
format_list_and(x)
}

\arguments{
\item{x}{list or vector (e.g., \code{letters[1:3]})}
}

\description{
Join a vector of items into a human-readable "a, b and c" list
}

