% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{format_year_range}
\alias{format_year_range}
\title{Format a year, or a range of years, as "2017" or "2017-2022"}
\usage{
format_year_range(a, b)
}

\arguments{
\item{a}{start year}

\item{b}{end year}
}

\description{
Format a year, or a range of years, as "2017" or "2017-2022"
}

