% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{growth_rate}
\alias{growth_rate}
\title{Compute Compound Annualized Growth Rate}
\usage{
growth_rate(p, q, t)
}

\arguments{
\item{p}{final value}

\item{q}{initial value}

\item{t}{time period}
}

\description{
Compute Compound Annualized Growth Rate
}

