% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in mod_cite.R
\name{mod_cite_server}
\alias{mod_cite_server}
\title{Citation server-side function}
\usage{
mod_cite_server(id)
}

\arguments{
\item{id}{Internal parameter for Shiny.}
}

\description{
Citation server-side function
}

