% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{od_treemap}
\alias{od_treemap}
\title{Origin-Destination Treemap}
\usage{
od_treemap(d, d2, title = NULL)
}

\arguments{
\item{d}{input dataset for values}

\item{d2}{input dataset for colours}

\item{title}{title for the treemap}
}

\description{
Origin-Destination Treemap
}

