% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{r_add_the}
\alias{r_add_the}
\title{Add definite article for reporter names}
\usage{
r_add_the(name = NULL)
}

\arguments{
\item{name}{Character string of the reporter name}
}

\value{
Character string: "the" for names requiring the article, empty string otherwise
}

\description{
Grammar helper function that adds "the" for reporter names such as
"United Kingdom" and "United States"
}

