% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{reactives}
\alias{rv}
\alias{rvtl}
\title{Typing reactiveValues is too long}
\usage{
rv(...)
rvtl(...)
}

\arguments{
\item{...}{elements to pass to the function}
}

\description{
Typing reactiveValues is too long
}

