% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{se_aggregate_by_sector}
\alias{se_aggregate_by_sector}
\title{Add Percentages to Sections}
\usage{
se_aggregate_by_sector(d, col, con)
}

\arguments{
\item{d}{input dataset}

\item{col}{column to collapse}

\item{con}{SQL connection}
}

\description{
Add Percentages to Sections
}

