% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{se_aggregate_sectors}
\alias{se_aggregate_sectors}
\title{Aggregate Sectors}
\usage{
se_aggregate_sectors(d, con)
}

\arguments{
\item{d}{input dataset}

\item{con}{SQL connection}
}

\description{
Aggregate Sectors
}

