% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils_server.R
\name{se_treemap}
\alias{se_treemap}
\title{Sector Treemap}
\usage{
se_treemap(d, d2, title = NULL)
}

\arguments{
\item{d}{input dataset for values}

\item{d2}{input dataset for colours}

\item{title}{title for the treemap}
}

\description{
Sector Treemap
}

