% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in tradestatisticsshiny-package.R
\name{tradestatisticsdashboard-package}
\alias{tradestatisticsdashboard-package}
\title{Open Trade Statistics Dashboard}
\description{
Iteractive dashboard to explore Open Trade Statistics data, which consists in a
 curated version of the International Trade and Production Database for Estimation (ITPD-E) and
 the International Trade and Production Database for Simulation (ITPD-S).
}

